requirements.txt
torch
torchvision
opencv-python
Pillow
tqdm
matplotlib
networkx
scipy
face_alignment
